@@ -20,18 +20,20 @@
         }
     ],
     "require": {
-        "php": "^8.1"
+        "php": "^8.2"
     },
     "require-dev": {
         "phpunit/phpunit": "^11",
         "monolog/monolog": "2.*",
         "phpstan/phpstan": "^1.10",
         "captainhook/captainhook": "^5.16",
-        "squizlabs/php_codesniffer": "^3.9"
+        "squizlabs/php_codesniffer": "^3.9",
+        "rector/rector": "^1.0"
     },
     "scripts": {
         "check-static": "phpstan analyse src --level=9",
         "format": "phpcbf --standard=PSR12 src tests",
+        "rector": "rector",
         "test": "phpunit tests",
         "post-autoload-dump": [
             "vendor/bin/captainhook install -f -s"
