@@ -8,24 +8,20 @@ namespace Devdot\Monolog;
  * @author Thomas Kuschan
  * @copyright (c) 2023
  */
-class LogRecord implements \ArrayAccess
+readonly class LogRecord implements \ArrayAccess
 {
     /**
      * Create a LogRecord to hold a single log entry.
-     * @param \DateTimeImmutable $datetime
-     * @param string $channel
-     * @param string $level
-     * @param string $message
      * @param array<int, mixed>|\stdClass|string|\NULL $context
      * @param array<int, mixed>|\stdClass|string|\NULL $extra
      */
     public function __construct(
-        public readonly \DateTimeImmutable $datetime,
-        public readonly string $channel,
-        public readonly string $level,
-        public readonly string $message,
-        public readonly array|\stdClass|string|null $context = [],
-        public readonly array|\stdClass|string|null $extra = [],
+        public \DateTimeImmutable $datetime,
+        public string $channel,
+        public string $level,
+        public string $message,
+        public array|\stdClass|string|null $context = [],
+        public array|\stdClass|string|null $extra = [],
     ) {
     }
 
