@@ -5,7 +5,7 @@ use Devdot\Monolog\LogRecord;
 
 final class LogRecordTest extends TestCase
 {
-    public function testConstruct()
+    public function testConstruct(): void
     {
         $record = new LogRecord(
             new \DateTimeImmutable('2023-01-02 08:00:01'),
@@ -26,7 +26,7 @@ final class LogRecordTest extends TestCase
         $this->assertCount(0, $record->extra);
     }
 
-    public function testArrayAccess()
+    public function testArrayAccess(): void
     {
         $record = new LogRecord(
             new \DateTimeImmutable('2023-01-02 08:00:01'),
@@ -56,5 +56,40 @@ final class LogRecordTest extends TestCase
         } catch (\Exception $e) {
             $this->assertInstanceOf(\LogicException::class, $e);
         }
+    }
+
+    public function testClassIsImmutable(): void
+    {
+        $date = new \DateTimeImmutable();
+        $record = new LogRecord(
+            $date,
+            'channel',
+            'level',
+            'hey message',
+        );
+        $this->assertSame($date, $record->datetime);
+        $this->assertSame('channel', $record->channel);
+        $this->assertSame('level', $record->level);
+        $this->assertSame('hey message', $record->message);
+        $this->assertSame([], $record->context);
+        $this->assertSame([], $record->extra);
+
+        $this->assertReadonly($record, 'datetime', new \DateTimeImmutable());
+        $this->assertReadonly($record, 'channel', '2');
+        $this->assertReadonly($record, 'level', 'up');
+        $this->assertReadonly($record, 'context', []);
+        $this->assertReadonly($record, 'extra', []);
+    }
+
+    private function assertReadonly(LogRecord $record, string $property, mixed $value = null): void
+    {
+        try {
+            $record->$property = $value;
+        } catch (\Error $e) {
+            if ($e->getMessage() === 'Cannot modify readonly property ' . $record::class . '::$' . $property) {
+                return;
+            }
+        }
+        $this->assertFalse(true, 'Property ' . $property . ' of ' . $record::class . ' is not readonly!');
     }
 }
